Project 1/requirements.txt
ccxt==4.1.77
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
plotly==5.17.0
python-dotenv==1.0.0 